requirements.txt
# ============================================================
# AgHealth+ / PAAI — Python dependencies
# Tested on Python 3.10 and 3.11
# ============================================================

# Core scientific stack
numpy==1.26.4
pandas==2.2.1
scipy==1.13.0
scikit-learn==1.4.2

# Reinforcement Learning
stable-baselines3==2.3.2
sb3-contrib==2.3.0          # MaskablePPO for constrained action spaces
gymnasium==0.29.1
torch==2.2.2                # CPU version sufficient for this workload
tensorboard==2.16.2

# Knowledge Graph
rdflib==7.0.0
networkx==3.3

# Data storage and serialisation
pyarrow==15.0.2             # Parquet long-term storage
python-dotenv==1.0.1
pyyaml==6.0.1

# Visualisation
matplotlib==3.8.4
seaborn==0.13.2

# MIMIC-IV access (PostgreSQL interface)
psycopg2-binary==2.9.9
sqlalchemy==2.0.29

# Cryptography (CIP audit log)
cryptography==42.0.5

# Statistical tests (DeLong AUC, Wilcoxon, effect sizes)
statsmodels==0.14.2
pingouin==0.5.4

# Utilities
tqdm==4.66.2
joblib==1.4.0

# Jupyter notebooks
jupyter==1.0.0
ipykernel==6.29.4

# Testing
pytest==8.1.1
pytest-cov==5.0.0

# Code quality
black==24.3.0
flake8==7.0.0
